PR Feedback - rename `controllerConnectionState` to `upstreamConnectionState` to correctly represent the state - Move `monitorUpstreamConnectionState()` test to `controller_connection_test.go`
