+++ internal/proto/worker/health/v1/health_service.proto
@@ -13,5 +13,5 @@ message HealthInfo {
   string state = 1;
   google.protobuf.UInt32Value active_session_count = 2 [json_name = "active_connection_count"];
   map<string, uint32> session_connections = 3 [json_name = "session_connections"];
-  string controller_connection_state = 4 [json_name = "controller_connection_state"];
+  string upstream_connection_state = 4 [json_name = "upstream_connection_state"];
 }
